requirements.txt
torch>=1.10.0
torchvision>=0.11.0
numpy>=1.19.0
Pillow>=8.0.0
tqdm>=4.60.0
huggingface-hub>=0.17.0